+++ modules/examples/examples/serdes_examples.cpp
@@ -7,7 +7,7 @@
 #include <fmt/core.h>
 
 #include "hephaestus/examples/types/pose.h"
-#include "hephaestus/examples/types_protobuf/pose.h"  // NOLINT(misc-include-cleaner)
+#include "hephaestus/examples/types_proto/pose.h"  // NOLINT(misc-include-cleaner)
 #include "hephaestus/serdes/json.h"
 #include "hephaestus/serdes/serdes.h"
