+++ modules/examples/include/hephaestus/examples/types_proto/pose.h
