+++ modules/examples/include/hephaestus/examples/types_proto/sample.h
