+++ modules/examples/src/types_proto/geometry.cpp
@@ -2,4 +2,4 @@
 // Copyright (C) 2023-2024 HEPHAESTUS Contributors
 //=================================================================================================
 
-#include "hephaestus/examples/types_protobuf/geometry.h"  // NOLINT(misc-include-cleaner)
+#include "hephaestus/examples/types_proto/geometry.h"  // NOLINT(misc-include-cleaner)
