+++ modules/examples/tests/CMakeLists.txt
@@ -3,8 +3,8 @@
 #=================================================================================================
 
 define_module_test(
-        NAME types_protobuf_tests
-        SOURCES types_protobuf_tests.cpp helpers.h
+        NAME types_proto_tests
+        SOURCES types_proto_tests.cpp helpers.h
         PUBLIC_INCLUDE_PATHS
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
         PUBLIC_LINK_LIBS "")
